Made some changes to the over-all quality of life around the pages.
Added full session support along with the ability for each page to recognize and respond differently to an authorized users.

Added a 404 access denied page with a random duck generator for good measure.

Modified the navbar to do some incredibly fancy CSS with flask conditions, a bit proud of that one (Even though I'm the only one who'll notice this probably)

Finally got around to actually adding tables; found a solid bootstrap boilerplate that I can work with for handling the pagination in this context. (Makes it a lot easier than parsing out a remainder from our full index of items)

Next on the agenda is to leverage flask conditions and session data to actually poll dynamic table headers and items, then we proceed to leveraging the CRUD functions I've already built out for data modifications.